kiruba(equipment)

+++ Tancem/src/main/java/com/Tancem/PIS/Controller/PlantDepartmentController.java
@@ -8,7 +8,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/tancem/pis/plant_department")
@@ -18,26 +20,53 @@ public class PlantDepartmentController {
     private PlantDepartmentService plantDepartmentService;
 
     @GetMapping("/readall")
-    public ResponseEntity<List<PlantDepartment>> getAllPlantDepartments() {
+    public ResponseEntity<Map<String, Object>> getAllPlantDepartments() {
         List<PlantDepartment> plantDepartments = plantDepartmentService.getAllPlantDepartments();
-        return new ResponseEntity<>(plantDepartments, HttpStatus.OK);
+        Map<String, Object> response = new HashMap<>();
+        response.put("statusCode", HttpStatus.OK.value());
+        response.put("statusMessage", "Successfully retrieved all plant departments");
+        response.put("data", plantDepartments);
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     @GetMapping("/read/{id}")
-    public ResponseEntity<PlantDepartment> getPlantDepartmentById(@PathVariable Integer id) {
+    public ResponseEntity<Map<String, Object>> getPlantDepartmentById(@PathVariable Integer id) {
         PlantDepartment plantDepartment = plantDepartmentService.getPlantDepartmentById(id);
-        return new ResponseEntity<>(plantDepartment, HttpStatus.OK);
+        Map<String, Object> response = new HashMap<>();
+        if (plantDepartment != null) {
+            response.put("statusCode", HttpStatus.OK.value());
+            response.put("statusMessage", "Successfully retrieved plant department");
+            response.put("data", plantDepartment);
+            return new ResponseEntity<>(response, HttpStatus.OK);
+        } else {
+            response.put("statusCode", HttpStatus.NOT_FOUND.value());
+            response.put("statusMessage", "Plant department not found");
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping("/create")
-    public ResponseEntity<PlantDepartment> createPlantDepartment(@RequestBody PlantDepartment plantDepartment) {
+    public ResponseEntity<Map<String, Object>> createPlantDepartment(@RequestBody PlantDepartment plantDepartment) {
         PlantDepartment newPlantDepartment = plantDepartmentService.savePlantDepartment(plantDepartment);
-        return new ResponseEntity<>(newPlantDepartment, HttpStatus.CREATED);
+        Map<String, Object> response = new HashMap<>();
+        response.put("statusCode", HttpStatus.CREATED.value());
+        response.put("statusMessage", "Plant department successfully created");
+        response.put("data", newPlantDepartment);
+        return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> deletePlantDepartment(@PathVariable Integer id) {
-        plantDepartmentService.deletePlantDepartment(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<Map<String, Object>> deletePlantDepartment(@PathVariable Integer id) {
+        Map<String, Object> response = new HashMap<>();
+        if (plantDepartmentService.getPlantDepartmentById(id) != null) {
+            plantDepartmentService.deletePlantDepartment(id);
+            response.put("statusCode", HttpStatus.NO_CONTENT.value());
+            response.put("statusMessage", "Plant department successfully deleted");
+            return new ResponseEntity<>(response, HttpStatus.NO_CONTENT);
+        } else {
+            response.put("statusCode", HttpStatus.NOT_FOUND.value());
+            response.put("statusMessage", "Plant department not found");
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
     }
 }
